pyproject.toml
[tool.poetry]
name = "lifegame"
version = "0.1.0"
description = ""
authors = ["NickiHell <nickihell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "*"
numpy = "*"
loguru = "*"
gym = "*"
pandas = "*"
redis = "*"

[tool.poetry.dev-dependencies]
ipython = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
